requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
google-generativeai>=0.8.0
openai>=1.30.0
tavily-python>=0.3.0
plotly>=5.18.0
